Rename icon component file. Make type required and active boolean
