+++ src/css/_icon.css
@@ -1,0 +1,5 @@
+/* Works with Black Tie icons */
+.icon-active {
+    color: yellow;
+    font-weight: 700;
+}
