+++ src/js/components/widgets-link/widget-link-vec2.react.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import Modal from 'react-bootstrap/lib/Modal';
 import Button from 'react-bootstrap/lib/Button';
 import DraggableModal from '../draggable-modal.react';
-import Icon from '../icon.react';
+import Icon from '../Icon';
 
 import Vector from './vector';
 import { editor } from '../../editor/editor';
