+++ src/js/components/widgets/widget-vector/widget-vector.react.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Modal from 'react-bootstrap/lib/Modal';
 import Button from 'react-bootstrap/lib/Button';
 import DraggableModal from '../../draggable-modal.react.js';
-import Icon from '../../icon.react';
+import Icon from '../../Icon';
 import { setCodeMirrorValue } from '../../../editor/editor';
 
 import THREE from 'three';
